Organize playground files
- Move Playground code into playground dir
- Break component sections into files
  e.g. Button, Checkbox, and Range sections of playground are broken into separate files
