+++ src/implementations/react/src/playground/PlaygroundSection.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+function PlaygroundSection(props) {
+  return (
+    <section>
+      <h3>{props.title}</h3>
+      {props.children}
+    </section>
+  )
+}
+
+PlaygroundSection.propTypes = {
+  title: PropTypes.string.isRequired,
+  children: PropTypes.node.isRequired
+}
+
+export default PlaygroundSection;
