+++ src/implementations/react/src/playground/images/profileImage.png
